process: add a way to allow event delivery to be delayed to a thread
It's possible with the InputManager event in particular for a thread to receive an event while it's holding a mutex that's also needed for the userspace event handler. The mutual exclusion works, leading to the two threads deadlocking (and bricking user input). This allows for event delivery to be delayed until after a particular block of code completes, which can be used to avoid this problem altogether.