ttyterm: inhibit input events while we read/write master pty
This resolves a bug where these paths lead to Buffer which takes a Mutex. Normally this is fine, but if the user presses a key while the kernel is processing one of these master pty read/writes, it ends up deadlocking on the mutex as the other holder is in the previous nesting level.