tui: move most of the TUI behavior into a library
I'm going to use this to help generalise the TUI code, which will allow me to have TUI work as an SVGA large-screen, single terminal, or as a windowed terminal in the window manager.

That would allow me to remove the window manager from releases but still have a pango-enabled much friendlier terminal available.