vfs: add a very basic first cut at reparse points
These are not at all integrated beyond the implementation in traversal and being able to set them.

These may be able to be used in the future to provide a UNIX-like filesystem (e.g. unixfs»/{bin,var,usr,dev}) by creating a structure full of directories with reparse points.

Worth noting: there's a lot more work to go here to figure out how best to have a good coexistence.