posix: make sure the target's signal handlers are used when sending signals
This fixes a bug where a receiver might receive an event from a sender, after which the sender quits. If the receiver doesn't get scheduled or otherwise check event state (e.g. acquiring a lock), the event gets destroyed but remains referenced by the Thread object. Then, when event state is checked, the Event gets pulled out of the queue and triggers a pure virtual function call.

This fixes that, and also makes sure the *target's* signal handler is the one that runs when receiving a signal, *NOT* the sender's.