kernel/process: acquire subsystem mutual exclusion in Process::~Process
Doing so ensures the Subsystem is free of other threads making changes before acquiring the Process lock. If this is done after the Process lock is acquired, it's possible that the locks are taken and the system will attempt to reschedule out of a held Spinlock critical section (which is not allowed).