kernel/process: if we handle events in PerProcessorScheduler::sleep, don't continue sleeping
This fixes a race in which other threads could run during the time the events are being handled.

One of the key examples of that race was a situation where a SIGCHLD event would be handled, and meanwhile the child that is terminating would run and release the waitpid notify lock. However, the sleep() continued after handling SIGCHLD, and so the lock never got checked again (even though it was unlocked). Having the scheduler abort the sleep if an event is handled resolved this.