kernel/process: add another case for switching to idle thread
In this case, when a thread is moved from one processor to another, but the original processor doesn't have any other threads, that processor needs to run the idle thread. Previously, it would run the original thread again and abort the reschedule.

This makes moving threads between cores more reliable rather than having it happen during Semaphore::acquire or similar.