processor/x86_common: create page stacks faster
This tries to do the writes to stacks in as tight a loop as possible, and also makes a few changes such as mapping the entire stacks first rather than mapping them on demand.

The best improvement would come from not doing the x64 stack initialization synchronously but rather doing that in a thread, allowing other tasks to use the physical page allocator without necessarily having access to pages above the 4 GB mark yet.