machine: add a trace mechanism
This will write basic log messages to a output mechanism that needs extremely minimal configuration such that traces can be used to track the progress of a kernel startup without needing to use the actual kernel log.

This allows for tracing global constructors and other early-boot processes, especially if/when they are hanging or otherwise breaking for an unknown reason.