kernel/utilities: use Result<> for RadixTree lookups
Also, I fixed a bug in RadixTree::lookup that was silently returning an empty value previously, as in some cases a node could split and then fail to have a value assigned to it later on. This affected a few things unexpectedly after adding in the Result<> usage as a genuine error state is now being used instead of just a falseish value.