x64: avoid race condition in VirtualAddressSpace
It's possible for a process to terminate while adjusting all other processes and in that case we can break sometimes by trying to access a process that no longer exists in the scheduler. For now, this is just being more defensive, but moving forward there should be a better way to iterate through the process tree safely.