posix: fix signedness issue with ioctls
The Linux definition of ioctl takes an unsigned long for the 'command'/'request' parameter - and Pedigree has been using an int. For some ioctls (e.g. those above 0x80000000) this breaks as the cast from int to size_t sign-extends and as a result the ioctl call doesn't match any known commands.

Using an unsigned type end-to-end fixes this up.