kernel: adjust RangeList to improve average performance
Now, allocate() and allocateSpecific() are the only places that call sweep(), and only if an initial allocation attempt fails. This defers merges until they're actually necessary which is convenient. Also, the main data type is now a Vector, which is far better for the kind of iteration that happens inside RangeList (List is fast to push to front or back, but not as good for iteration).