kernel: add Cord/StaticCord data types and integrate into Log
Cords are useful as they are basically compositions of disparate string components without requiring creation of a String or similar object. They are much faster to generate (~20% performance boost in Log::flushEntry).

Still need to add proper iteration that keeps track of where in the Cord the iteration is so things like SerialLogger can more efficiently traverse the cord.