kernel: update SymbolTable to better track parented weak and global symbols
This was needed as eraseByElf() was taking quite some time in startup on failed modules, trying to remove the failed module's symbols from the global/weak symbol tables. In many cases this was removing dozens of symbols from a table holding over 100k symbols, which is not particularly efficient to do (it required iterating over all 100k symbols, removing in-flight, which may have caused hash table resizes too).