kernel: start using less preprocessor conditions
There's a lot more work to go before this is really "done", but right now a lot more code is getting checked in each compile. Previously, toggling flags in the build system would cause a number of issues due to code paths being enabled that hadn't been tested for a while. Now much less should be broken because much more of the kernel is compiled each time (even if it's not necessarily emitted as it's behind a constant branch).